requirements.txt
Appium-Python-Client == 3.1.1   
attrs == 23.2.0     
certifi == 2024.2.2   
exceptiongroup == 1.2.0      
h11 == 0.14.0     
idna == 3.6        
iniconfig == 2.0.0      
outcome == 1.3.0.post0
packaging == 24.0       
pip == 19.2.3     
pluggy == 1.4.0      
pyclean == 3.0.0      
PySocks == 1.7.1      
pytest == 8.1.1      
selenium == 4.18.1     
setuptools == 41.2.0     
sniffio == 1.3.1      
sortedcontainers == 2.4.0      
tomli == 2.0.1      
trio == 0.25.0
trio-websocket == 0.11.1
typing-extensions == 4.10.0
urllib3 == 2.2.1     
wsproto == 1.2.0